Move pagination logic into separate store
The primary store still contains a copy of pagination state
that is updated after each pagination action. This is due
to the fact that data from the primary store is
automatically updated in components. Presently  it is necessary
that the values exist in the primary store so that the
components can access them.