Add display of total page count
The display of total pages is hidden by default (clicking the current
page number will toggle visibility). This was done because the estimated
total number of pages, while calculated, is never reliable and is often
misleading. More info will be added to the README.

The item module's prefetch will now trigger a hard-setting of the last
page when a prefetch request returns no items. This offers the most
reliable source of page count, but often involves the totalPages display
changing from > 100 to ~20.